requirements.txt
torch
fair-esm
gymnasium
git  https://github.com/Superzchen/iFeature.git #need to clone inside src folder
modlamp
pymc
seaborn
wandb
scipy
scikit-learn
tqdm
easydict
pyyaml
stable-baselines3